requirements.txt
asgiref==3.2.7
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.39.1
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
docutils==0.16
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
psycopg2-binary
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==3.3.1
wrapt==1.11.2
django-colorfield
django-import-export
django-redis==4.12.1